0.5.0

+++ lib/cli/cli.go
@@ -2,10 +2,10 @@ package cli
 
 import (
 	"fmt"
-	"log"
 	"strconv"
 	"time"
 
+	log "github.com/Sirupsen/logrus"
 	"github.com/fatih/color"
 	"github.com/olekukonko/tablewriter"
 	"github.com/struCoder/pmgo/lib/master"
@@ -38,15 +38,6 @@ func (cli *Cli) Save() {
 	}
 }
 
-// Resurrect will restore all previously save processes.
-// Display an error in case there's any.
-func (cli *Cli) Resurrect() {
-	err := cli.remoteClient.Resurrect()
-	if err != nil {
-		log.Fatalf("Failed to resurrect all previously save processes due to: %+v\n", err)
-	}
-}
-
 // StartGoBin will try to start a go binary process.
 // Returns a fatal error in case there's any.
 func (cli *Cli) StartGoBin(sourcePath string, name string, keepAlive bool, args []string) {
@@ -59,6 +50,11 @@ func (cli *Cli) StartGoBin(sourcePath string, name string, keepAlive bool, args 
 // RestartProcess will try to restart a process with procName. Note that this process
 // must have been already started through StartGoBin.
 func (cli *Cli) RestartProcess(procName string) {
+	isExist := cli.remoteClient.GetProcByName(procName)
+	if len(*isExist) == 0 {
+		log.Errorf("porcess %s not found", procName)
+		return
+	}
 	err := cli.remoteClient.RestartProcess(procName)
 	if err != nil {
 		log.Fatalf("Failed to restart process due to: %+v\n", err)
@@ -70,20 +66,30 @@ func (cli *Cli) RestartProcess(procName string) {
 func (cli *Cli) StartProcess(procName string) {
 	err := cli.remoteClient.StartProcess(procName)
 	if err != nil {
-		log.Fatalf("Failed to start process due to: %+v\n", err)
+		log.Errorf("Failed to start process due to: %+v\n", err)
 	}
 }
 
 // StopProcess will try to stop a process named procName.
 func (cli *Cli) StopProcess(procName string) {
-	err := cli.remoteClient.StopProcess(procName)
-	if err != nil {
-		log.Fatalf("Failed to stop process due to: %+v\n", err)
+	isExist := cli.remoteClient.GetProcByName(procName)
+	if len(*isExist) == 0 {
+		log.Warnf("porcess %s not found", procName)
+	} else {
+		err := cli.remoteClient.StopProcess(procName)
+		if err != nil {
+			log.Fatalf("Failed to stop process due to: %+v\n", err)
+		}
 	}
 }
 
 // DeleteProcess will stop and delete all dependencies from process procName forever.
 func (cli *Cli) DeleteProcess(procName string) {
+	isExist := cli.remoteClient.GetProcByName(procName)
+	if len(*isExist) == 0 {
+		log.Errorf("porcess %s not found", procName)
+		return
+	}
 	err := cli.remoteClient.DeleteProcess(procName)
 	if err != nil {
 		log.Fatalf("Failed to delete process due to: %+v\n", err)
@@ -122,15 +128,36 @@ func (cli *Cli) Status() {
 
 // ProcInfo will display process information
 func (cli *Cli) ProcInfo(procName string) {
+	procDetail := cli.remoteClient.GetProcByName(procName)
+	if len(*procDetail) == 0 {
+		log.Errorf("porcess %s not found", procName)
+		return
+	}
 	table := utils.GetTableWriter()
 	table.SetAutoWrapText(true)
 	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	procDetail := cli.remoteClient.GetProcByName(procName)
 	for k, v := range *procDetail {
 		table.Append([]string{
 			color.GreenString(k), v,
 		})
 	}
 	table.Render()
+}
 
+// DeleteAllProcess will stop all process
+func (cli *Cli) DeleteAllProcess() {
+	procResponse, err := cli.remoteClient.MonitStatus()
+	if err != nil {
+		log.Fatalf("Failed to get status due to: %+v\n", err)
+	}
+
+	if len(procResponse.Procs) == 0 {
+		log.Warn("All processes have been stopped and deleted")
+		return
+	}
+	for id := range procResponse.Procs {
+		proc := procResponse.Procs[id]
+		cli.remoteClient.DeleteProcess(proc.Name)
+		log.Infof("proc: %s has quit", proc.Name)
+	}
 }
